Feature/22 add logi

See merge request simbirsoft1/autumn-training-2021/homeowners-data-app/backend!24

+++ src/main/java/com/simbirsoftintensiv/intensiv/controller/user/AdminRestController.java
@@ -1,28 +1,27 @@
 package com.simbirsoftintensiv.intensiv.controller.user;
 
+import com.simbirsoftintensiv.intensiv.controller.LoginController;
+import com.simbirsoftintensiv.intensiv.entity.User;
+import com.simbirsoftintensiv.intensiv.exception_handling.NoSuchUserException;
 import com.simbirsoftintensiv.intensiv.service.user.UserService;
 import com.simbirsoftintensiv.intensiv.to.UserTo;
 import com.simbirsoftintensiv.intensiv.util.UserUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.stream.Collectors;
 
-/*
-После авторизации Юзера и если у него роль ADMIN
-Фронт делает GET запрос на /admin
-и ему приходит список/json со списком юзеров в БД
-при POST запросе на /admin?action=delete&userId={userId}
-происходит удаление
- */
-
 
 @RestController
 @RequestMapping(value = "/rest/admin/users")
 public class AdminRestController {
-
+    static final Logger log =
+            LoggerFactory.getLogger(LoginController.class);
     private final UserService userService;
 
     public AdminRestController(UserService userService) {
@@ -31,79 +30,47 @@ public class AdminRestController {
 
     @GetMapping
     public List<UserTo> getAll() {
+        log.info("Админ получил данные на всех пользователей.");
+//TODO а если несколько админов то надо будет сделать какой конкретно
         return userService.getAll()
                 .stream()
                 .map(UserUtil::asTo)
                 .collect(Collectors.toList());
     }
-   /* @ResponseBody
-    @GetMapping
-    public HashMap<String, List<User>> getAll() {
-        System.out.println("создаем массив с логинами юзеров");
-        HashMap<String, List<User>> map = new HashMap<>();
-        map.put("Content", userService.getAll());
-        return map;
-    }*/
+
 
     @DeleteMapping("/{phone}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable("phone") long phone) {
+        User user = userService.getByPhone(phone);
+        if (user == null) {
+            log.warn("Попытка удаление не существующего пользователя.");
+            throw new NoSuchUserException("User" +phone + " not find ");
+        }
+        log.warn("User " + phone + "deleting.");
+
         userService.delete(phone);
     }
-/*
-    @PostMapping("/admin")
-    public HashMap<String, String> deleteUser(@RequestParam(required = true, defaultValue = "") Integer userId,
-                                              @RequestParam(defaultValue = "") String action) {
 
-        HashMap<String, String> map = new HashMap<>();
-        if (action.equals("DELETE")) {
-            boolean itDelete = userService.delete(userId);
-            if (itDelete) {
-                map.put("code", "200 OK");
-            } else {
-                map.put("code", "404");
-            }
-        }
-        return map;
-    }
-*/
     @GetMapping("/{phone}")
     public UserTo get(@PathVariable("phone") long phone) {
+        User user = userService.getByPhone(phone);
+        if (user == null) {
+            log.warn("Попытка получение данных не существующего пользователя.");
+            throw new NoSuchUserException("User" +phone + " not find ");
+        }
+        log.info("Админ получил данные на пользователя " + phone);
         return UserUtil.asTo(userService.getByPhone(phone));
     }
-/*
-    @GetMapping("/user/{phone}")
-    public HashMap gtUser(@PathVariable("userId") String userId) {
-        Integer rightId;
-        List<String> code;
-        HashMap<String, List> map = new HashMap<>();
-        try {
-            rightId = Integer.parseInt(userId);
-        } catch (NumberFormatException exception) {
-            code = Collections.singletonList("400");
-            map.put("code", code);
-            return map;
-        }
-        List User = userService.usergtList(rightId);
-
-        if (getAll().size() == 1) {
-            code = Collections.singletonList("404");
-        } else {
-            System.out.println(getAll().size());
-            code = Collections.singletonList("200");
-        }
-        map.put("Content", User);
-        map.put("code", code);
-        return map;
-    }*/
 
     @ResponseBody
-    @GetMapping("/admin/count")
+    @GetMapping("/count")
     public HashMap<String, Integer> getUsersCount() {
         HashMap<String, Integer> map = new HashMap<>();
         Integer countUsers = userService.getAll().size();
-        map.put("code", 200);
-        map.put("Content", countUsers);
+        map.put("allUsers", countUsers);
+        log.info("Админ запрашивал кол-во пользователей");
+
         return map;
     }
